+++ AndroidStudioProject/app/src/main/java/org/gkisalatiga/plus/screen/ScreenAbout.kt
@@ -11,27 +11,43 @@ package org.gkisalatiga.plus.screen
 
 import android.annotation.SuppressLint
 import android.content.Context
+import android.content.Intent
 import android.content.pm.ApplicationInfo
 import android.content.pm.PackageInfo
+import android.net.Uri
 import android.util.Log
 import android.widget.Toast
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.BackHandler
 import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material.icons.automirrored.filled.ForwardToInbox
+import androidx.compose.material.icons.automirrored.filled.LibraryBooks
+import androidx.compose.material.icons.filled.Badge
+import androidx.compose.material.icons.filled.Code
+import androidx.compose.material.icons.filled.Diversity1
+import androidx.compose.material.icons.filled.History
+import androidx.compose.material.icons.filled.OpenInBrowser
+import androidx.compose.material.icons.filled.QuestionMark
+import androidx.compose.material.icons.filled.Security
+import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
@@ -40,28 +56,46 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.material3.TextField
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.material3.rememberTopAppBarState
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.LocalUriHandler
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import androidx.compose.ui.window.Dialog
+import com.halilibo.richtext.commonmark.Markdown
+import com.halilibo.richtext.ui.material.RichText
 import org.gkisalatiga.plus.R
 import org.gkisalatiga.plus.global.GlobalSchema
 import org.gkisalatiga.plus.lib.NavigationRoutes
-import org.gkisalatiga.plus.services.AlarmService
 import org.gkisalatiga.plus.services.NotificationService
 import java.io.File
+import java.io.InputStream
 
 
-class ScreenAbout() : ComponentActivity() {
+class ScreenAbout : ComponentActivity() {
+
+    // For displaying the license dialogs.
+    private val showLicenseDialog = mutableStateOf(false)
+
+    // For displaying generic markdown dialog.
+    private val showMarkdownDialog = mutableStateOf(false)
+    private var dialogMarkdownTitle = ""
+    private var dialogMarkdownContent = ""
+    private var dialogMarkdownIcon = Icons.Default.QuestionMark
 
     @Composable
     @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
@@ -86,7 +120,7 @@ class ScreenAbout() : ComponentActivity() {
             topBar = { getTopBar() }
                 ) {
 
-            val scrollState = rememberScrollState()
+            val scrollState = GlobalSchema.screenAboutScrollState!!
             Column (
                 verticalArrangement = Arrangement.Top,
                 horizontalAlignment = Alignment.CenterHorizontally,
@@ -95,7 +129,7 @@ class ScreenAbout() : ComponentActivity() {
                     .fillMaxSize()
                     .verticalScroll(scrollState)) {
                 /* Show app logo, name, and version. */
-                Box (Modifier.padding(vertical = 15.dp)) {
+                Box (Modifier.padding(vertical = 15.dp).padding(top = 10.dp)) {
                     Surface (
                         shape = CircleShape,
                         modifier = Modifier.size(100.dp),
@@ -128,11 +162,20 @@ class ScreenAbout() : ComponentActivity() {
                     }
                 }
                 Text(appName, fontWeight = FontWeight.Bold, fontSize = 24.sp)
-                Text(vName, fontSize = 18.sp)
-
+                Text("Versi $vName", fontSize = 18.sp)
                 Spacer(Modifier.height(20.dp))
-                Text("Dibuat oleh Sam, Jeffrey, dan Joaquim")
-                Text("Semoga di-approve sama MJ \uD83D\uDDFF :v")
+                Text(stringResource(R.string.app_description), modifier = Modifier.padding(horizontal = 20.dp), textAlign = TextAlign.Center)
+                Spacer(Modifier.height(20.dp))
+
+                // Display the main "about" contents.
+                getMainContent()
+
+                // Trailing space for visual improvement.
+                Spacer(Modifier.height(20.dp))
+
+                // Displaying the about dialogs.
+                getDialogLicense()
+                getDialogMarkdown()
             }
 
         }
@@ -144,6 +187,159 @@ class ScreenAbout() : ComponentActivity() {
             GlobalSchema.pushScreen.value = NavigationRoutes().SCREEN_MAIN
         }
 
+    }
+
+    @Composable
+    private fun getMainContent() {
+        getAppInfo()
+        Spacer(Modifier.height(20.dp))
+        getAuthorInfo()
+    }
+
+    @Composable
+    private fun getAppInfo() {
+        val ctx = LocalContext.current
+        val uriHandler = LocalUriHandler.current
+
+        /* Section: App Info */
+        Column (Modifier.fillMaxWidth()) {
+            val appInfoText = stringResource(R.string.screen_about_tentang_aplikasi)
+            Spacer(Modifier.height(10.dp))
+            Text(appInfoText, modifier = Modifier.padding(start = 20.dp), fontWeight = FontWeight.Bold, fontSize = 20.sp, overflow = TextOverflow.Ellipsis)
+            Spacer(Modifier.height(10.dp))
+
+            /* App License. */
+            val licenseText = stringResource(R.string.screen_about_lisensi)
+            Surface(
+                modifier = Modifier.fillMaxWidth().padding(0.dp).height(50.dp),
+                onClick = { showLicenseDialog.value = true }
+            ) {
+                Row (modifier = Modifier.padding(horizontal = 10.dp), verticalAlignment = Alignment.CenterVertically) {
+                    Icon(Icons.Default.Badge, "", modifier = Modifier.fillMaxHeight().padding(horizontal = 20.dp))
+                    Text(licenseText, modifier = Modifier, textAlign = TextAlign.Center)
+                }
+            }
+
+            /* App privacy policy. */
+            val privacyPolicyText = stringResource(R.string.screen_about_kebijakan_privasi)
+            Surface(
+                modifier = Modifier.fillMaxWidth().padding(0.dp).height(50.dp),
+                onClick = {
+                    // Prepare the text from raw resource.
+                    val input: InputStream = ctx.resources.openRawResource(R.raw.app_privacy_policy_en)
+                    val inputAsString: String = input.bufferedReader().use { it.readText() }
+
+                    // Displaying the text.
+                    dialogMarkdownContent = inputAsString
+                    dialogMarkdownIcon = Icons.Default.Security
+                    dialogMarkdownTitle = privacyPolicyText
+                    showMarkdownDialog.value = true
+                }
+            ) {
+                Row (modifier = Modifier.padding(horizontal = 10.dp), verticalAlignment = Alignment.CenterVertically) {
+                    Icon(Icons.Default.Security, "", modifier = Modifier.fillMaxHeight().padding(horizontal = 20.dp))
+                    Text(privacyPolicyText, modifier = Modifier, textAlign = TextAlign.Center)
+                }
+            }
+
+            /* App attribution and third party libraries. */
+            val attributionText = stringResource(R.string.screen_about_atribusi_pihak_ketiga)
+            Surface(
+                modifier = Modifier.fillMaxWidth().padding(0.dp).height(50.dp),
+                onClick = {
+                    GlobalSchema.pushScreen.value = NavigationRoutes().SCREEN_ATTRIBUTION
+                }
+            ) {
+                Row (modifier = Modifier.padding(horizontal = 10.dp), verticalAlignment = Alignment.CenterVertically) {
+                    Icon(Icons.AutoMirrored.Default.LibraryBooks, "", modifier = Modifier.fillMaxHeight().padding(horizontal = 20.dp))
+                    Text(attributionText, modifier = Modifier, textAlign = TextAlign.Center)
+                }
+            }
+
+            /* App source code. */
+            val sourceCodeText = stringResource(R.string.screen_about_kode_sumber)
+            Surface(
+                modifier = Modifier.fillMaxWidth().padding(0.dp).height(50.dp),
+                onClick = { uriHandler.openUri(GlobalSchema.aboutSourceCodeURL) }
+            ) {
+                Row (modifier = Modifier.padding(horizontal = 10.dp), verticalAlignment = Alignment.CenterVertically) {
+                    Icon(Icons.Default.Code, "", modifier = Modifier.fillMaxHeight().padding(horizontal = 20.dp))
+                    Text(sourceCodeText, modifier = Modifier, textAlign = TextAlign.Center)
+                    Icon(Icons.Default.OpenInBrowser,"", modifier = Modifier.padding(start = 5.dp).height(14.dp))
+                }
+            }
+
+            /* App source code. */
+            val changelogText = stringResource(R.string.screen_about_log_perubahan)
+            Surface(
+                modifier = Modifier.fillMaxWidth().padding(0.dp).height(50.dp),
+                onClick = { uriHandler.openUri(GlobalSchema.aboutChangelogURL) }
+            ) {
+                Row (modifier = Modifier.padding(horizontal = 10.dp), verticalAlignment = Alignment.CenterVertically) {
+                    Icon(Icons.Default.History, "", modifier = Modifier.fillMaxHeight().padding(horizontal = 20.dp))
+                    Text(changelogText, modifier = Modifier, textAlign = TextAlign.Center)
+                    Icon(Icons.Default.OpenInBrowser,"", modifier = Modifier.padding(start = 5.dp).height(14.dp))
+                }
+            }
+
+        }  // --- end of column: section app info.
+    }
+
+    @Composable
+    private fun getAuthorInfo() {
+        val ctx = LocalContext.current
+
+        /* Section: Author Info */
+        Column (Modifier.fillMaxWidth()) {
+            val appInfoText = stringResource(R.string.screen_about_tentang_pengembang)
+            Spacer(Modifier.height(10.dp))
+            Text(appInfoText, modifier = Modifier.padding(start = 20.dp), fontWeight = FontWeight.Bold, fontSize = 20.sp, overflow = TextOverflow.Ellipsis)
+            Spacer(Modifier.height(10.dp))
+
+            /* Contributor lists. */
+            val contributorText = stringResource(R.string.screen_about_kontributor)
+            Surface(
+                modifier = Modifier.fillMaxWidth().padding(0.dp).height(50.dp),
+                onClick = {
+                    // Prepare the text from raw resource.
+                    val input: InputStream = ctx.resources.openRawResource(R.raw.app_contribution)
+                    val inputAsString: String = input.bufferedReader().use { it.readText() }
+
+                    // Displaying the text.
+                    dialogMarkdownContent = inputAsString
+                    dialogMarkdownIcon = Icons.Default.Diversity1
+                    dialogMarkdownTitle = contributorText
+                    showMarkdownDialog.value = true
+                }
+            ) {
+                Row (modifier = Modifier.padding(horizontal = 10.dp), verticalAlignment = Alignment.CenterVertically) {
+                    Icon(Icons.Default.Diversity1, "", modifier = Modifier.fillMaxHeight().padding(horizontal = 20.dp))
+                    Text(contributorText, modifier = Modifier, textAlign = TextAlign.Center)
+                }
+            }
+
+            /* Author contact. */
+            val contactText = stringResource(R.string.screen_about_kontak)
+            val emailChooserTitle = stringResource(R.string.email_chooser_title)
+            Surface(
+                modifier = Modifier.fillMaxWidth().padding(0.dp).height(50.dp),
+                onClick = {
+                    // SOURCE: https://www.geeksforgeeks.org/how-to-send-an-email-from-your-android-app/
+                    // SOURCE: https://www.tutorialspoint.com/android/android_sending_email.htm
+                    // SOURCE: https://stackoverflow.com/a/59365539
+                    val emailIntent = Intent(Intent.ACTION_SENDTO)
+                    emailIntent.setData(Uri.parse("mailto:${GlobalSchema.aboutContactMail}"))
+                    ctx.startActivity(Intent.createChooser(emailIntent, emailChooserTitle))
+                }
+            ) {
+                Row (modifier = Modifier.padding(horizontal = 10.dp), verticalAlignment = Alignment.CenterVertically) {
+                    Icon(Icons.AutoMirrored.Default.ForwardToInbox, "", modifier = Modifier.fillMaxHeight().padding(horizontal = 20.dp))
+                    Text(contactText, modifier = Modifier, textAlign = TextAlign.Center)
+                    Icon(Icons.Default.OpenInBrowser,"", modifier = Modifier.padding(start = 5.dp).height(14.dp))
+                }
+            }
+
+        }  // --- end of column: section author info.
     }
 
     @Composable
@@ -176,4 +372,89 @@ class ScreenAbout() : ComponentActivity() {
             scrollBehavior = scrollBehavior
         )
     }
+
+    @Composable
+    private fun getDialogLicense() {
+        val ctx = LocalContext.current
+        val uriHandler = LocalUriHandler.current
+        val scrollState = rememberScrollState()
+
+        if (showLicenseDialog.value) {
+            AlertDialog(
+                icon = {
+                    Icon(Icons.Default.Badge, "")
+                },
+                title = {
+                    Text(stringResource(R.string.screen_about_lisensi))
+                },
+                text = {
+                    Box(Modifier.height(250.dp).background(Color.Transparent)) {
+                        // Prepare the license text from raw resource.
+                        val input: InputStream = ctx.resources.openRawResource(R.raw.app_license)
+                        val inputAsString: String = input.bufferedReader().use { it.readText() }
+
+                        // Display the license text.
+                        Column(Modifier.verticalScroll(scrollState).fillMaxSize()) {
+                            Text(inputAsString, modifier = Modifier.fillMaxSize())
+                        }
+                    }
+                },
+                onDismissRequest = {
+                    showLicenseDialog.value = false
+                },
+                confirmButton = {
+                    TextButton(onClick = { uriHandler.openUri(GlobalSchema.aboutLicenseFullTextURL) }) {
+                        Text(stringResource(R.string.screen_about_license_confirm))
+                    }
+                },
+                dismissButton = {
+                    TextButton(onClick = { showLicenseDialog.value = false }) {
+                        Text(stringResource(R.string.screen_about_license_dismiss))
+                    }
+                }
+            )
+
+        }
+    }
+
+    @Composable
+    private fun getDialogMarkdown() {
+        val ctx = LocalContext.current
+        val uriHandler = LocalUriHandler.current
+        val scrollState = rememberScrollState()
+
+        if (showMarkdownDialog.value) {
+            AlertDialog(
+                icon = {
+                    Icon(dialogMarkdownIcon, "")
+                },
+                title = {
+                    Text(dialogMarkdownTitle)
+                },
+                text = {
+                    Box(Modifier.height(250.dp).background(Color.Transparent)) {
+
+                        // Display the markdown text.
+                        Column(Modifier.verticalScroll(scrollState).fillMaxSize()) {
+                            RichText {
+                                Markdown(dialogMarkdownContent.trimIndent())
+                            }
+                        }
+
+                    }
+                },
+                onDismissRequest = {
+                    showMarkdownDialog.value = false
+                },
+                confirmButton = {},
+                dismissButton = {
+                    TextButton(onClick = { showMarkdownDialog.value = false }) {
+                        Text(stringResource(R.string.screen_about_license_dismiss))
+                    }
+                }
+            )
+
+        }
+    }
+
 }