+++ AndroidStudioProject/app/src/main/java/org/gkisalatiga/plus/screen/ScreenAttribution.kt
@@ -1,0 +1,192 @@
+/**
+ * AGPL-3.0-licensed
+ * Copyright (C) GKI Salatiga 2024
+ * Written by Samarthya Lykamanuella (github.com/groaking)
+ *
+ * ---
+ * Display meta-application information about GKI Salatiga Plus.
+ */
+
+package org.gkisalatiga.plus.screen
+
+import android.annotation.SuppressLint
+import android.content.Context
+import android.content.Intent
+import android.content.pm.ApplicationInfo
+import android.content.pm.PackageInfo
+import android.net.Uri
+import android.util.Log
+import android.widget.Toast
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.BackHandler
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.heightIn
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material.icons.automirrored.filled.ForwardToInbox
+import androidx.compose.material.icons.automirrored.filled.LibraryBooks
+import androidx.compose.material.icons.filled.Badge
+import androidx.compose.material.icons.filled.Code
+import androidx.compose.material.icons.filled.Diversity1
+import androidx.compose.material.icons.filled.History
+import androidx.compose.material.icons.filled.OpenInBrowser
+import androidx.compose.material.icons.filled.Security
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.CenterAlignedTopAppBar
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.material3.TextField
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.material3.rememberTopAppBarState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.LocalUriHandler
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.compose.ui.window.Dialog
+import org.gkisalatiga.plus.R
+import org.gkisalatiga.plus.global.GlobalSchema
+import org.gkisalatiga.plus.lib.AppDatabase
+import org.gkisalatiga.plus.lib.NavigationRoutes
+import org.gkisalatiga.plus.services.NotificationService
+import org.json.JSONArray
+import org.json.JSONObject
+import java.io.File
+import java.io.InputStream
+
+
+class ScreenAttribution : ComponentActivity() {
+
+    @Composable
+    @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
+    public fun getComposable() {
+
+        Log.d("Groaker", "Last selected fragment of main screen: ${GlobalSchema.lastMainScreenPagerPage.value}")
+
+        Scaffold (
+            topBar = { getTopBar() }
+                ) {
+
+            val scrollState = GlobalSchema.screenAttributionScrollState!!
+            Column (
+                verticalArrangement = Arrangement.Top,
+                horizontalAlignment = Alignment.CenterHorizontally,
+                modifier = Modifier
+                    .padding(top = it.calculateTopPadding(), bottom = it.calculateBottomPadding())
+                    .fillMaxSize()
+                    .verticalScroll(scrollState)) {
+                // Display the main "attribution" contents.
+                getMainContent()
+            }
+
+        }
+
+        // Ensure that when we are at the first screen upon clicking "back",
+        // the app is exited instead of continuing to navigate back to the previous screens.
+        // SOURCE: https://stackoverflow.com/a/69151539
+        BackHandler {
+            GlobalSchema.pushScreen.value = NavigationRoutes().SCREEN_ABOUT
+        }
+
+    }
+
+    @Composable
+    private fun getMainContent() {
+        // This is used to open external URLs.
+        val uriHandler = LocalUriHandler.current
+
+        // Load the JSON file containing attributions of all open source programs
+        // and code which are used by this app.
+        val attribJSON: JSONObject = AppDatabase().getAttributions()
+        val attribArray: JSONArray = attribJSON.getJSONArray("opensource-attributions")
+
+        // Convert JSONArray to regular list.
+        val attribList: MutableList<JSONObject> = mutableListOf()
+        for (i in 0 until attribArray.length()) {
+            attribList.add(attribArray[i] as JSONObject)
+        }
+
+        /* Displaying the attribution cards. */
+        attribList.forEach {
+
+            /* The attribution card. */
+            Surface (
+                onClick = { uriHandler.openUri(it.getString("link")) },
+                modifier = Modifier.fillMaxWidth()
+            ) {
+                Column (modifier = Modifier.fillMaxWidth().padding(20.dp).padding(top = 5.dp), verticalArrangement = Arrangement.Center) {
+                    Text(it.getString("title"), fontWeight = FontWeight.Bold)
+                    Text("Copyright (C) ${it.getString("year")} ${it.getString("author")}")
+                    TextButton(onClick = { uriHandler.openUri(it.getString("license-url")) }) {
+                        Text(it.getString("license"))
+                    }
+                }
+            }
+
+        }
+
+    }
+
+    @Composable
+    @OptIn(ExperimentalMaterial3Api::class)
+    private fun getTopBar() {
+        val scrollBehavior = TopAppBarDefaults.pinnedScrollBehavior(rememberTopAppBarState())
+        CenterAlignedTopAppBar(
+            colors = TopAppBarDefaults.centerAlignedTopAppBarColors(
+                containerColor = MaterialTheme.colorScheme.primaryContainer,
+                titleContentColor = MaterialTheme.colorScheme.primary
+            ),
+            title = {
+                Text(
+                    stringResource(R.string.screenattrib_title),
+                    maxLines = 1,
+                    overflow = TextOverflow.Ellipsis
+                )
+            },
+            navigationIcon = {
+                IconButton(onClick = {
+                    GlobalSchema.pushScreen.value = NavigationRoutes().SCREEN_ABOUT
+                }) {
+                    Icon(
+                        imageVector = Icons.AutoMirrored.Default.ArrowBack,
+                        contentDescription = ""
+                    )
+                }
+            },
+            actions = { },
+            scrollBehavior = scrollBehavior
+        )
+    }
+
+}